+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## HEAD
-
+- `bnsd`: when a domain is transferred accounts ownership is transferred to the new domain owner and accounts' targets are cleared
 - `bnsapi`: move bnsapi to new repo
 - `bnscli`: fix boolean flag bug
 - `bug`: Refactor register domain message [issue#1165](https://github.com/iov-one/weave/issues/1165)